[ci] Automated git mirroring from cr.imson.co repository

+++ ansible/roles/volume-backup/files/hooks/snapshot-backup/backup-repo.sh
@@ -13,17 +13,19 @@ set -o errexit -o pipefail -o noclobber -o nounset
 . /srv/homelab/homelab_config.sh
 . "$HOMELAB_CONFIG_PATH/backup_config.sh"
 
-echo :: backing up homelab repo...
+if [[ "$BACKUP_BACKUP_REPO" == "true" ]]; then
+    echo :: backing up homelab repo...
 
-TEMP_REPO_PATH=$(mktemp -d -p "$BACKUP_STAGING_PATH/")
+    TEMP_REPO_PATH=$(mktemp -d -p "$BACKUP_STAGING_PATH/")
 
-pushd "$TEMP_REPO_PATH/" > /dev/null
+    pushd "$TEMP_REPO_PATH/" > /dev/null
 
-git clone --depth 1 "$BACKUP_GIT_REPO" .
+    git clone --depth 1 "$BACKUP_GIT_REPO" .
 
-HOMELAB_REPO_SHA=$(git rev-parse --short HEAD)
-git archive --format=tar main | bzip2 -c > "$BACKUP_STAGING_PATH/homelab-repo.$HOMELAB_REPO_SHA.tar.bz2"
+    HOMELAB_REPO_SHA=$(git rev-parse --short HEAD)
+    git archive --format=tar main | bzip2 -c > "$BACKUP_STAGING_PATH/homelab-repo.$HOMELAB_REPO_SHA.tar.bz2"
 
-popd > /dev/null
+    popd > /dev/null
 
-rm -rf "${TEMP_REPO_PATH:?}/"
+    rm -rf "${TEMP_REPO_PATH:?}/"
+fi
